requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
django-anymail==10.1
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
